src/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.3
    # via deepgram-sdk
aioredis==1.3.1
    # via channels-redis
aiosignal==1.2.0
    # via aiohttp
amazon-transcribe==0.6.0
    # via -r requirements.in
anyio==3.6.1
    # via watchfiles
asgiref==3.5.2
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
async-timeout==4.0.2
    # via
    #   aiohttp
    #   aioredis
attrs==22.1.0
    # via
    #   aiohttp
    #   automat
    #   jsonschema
    #   service-identity
    #   twisted
autobahn==22.7.1
    # via daphne
automat==20.2.0
    # via twisted
awscrt==0.13.14
    # via amazon-transcribe
azure-cognitiveservices-speech==1.23.0
    # via -r requirements.in
cachetools==5.2.0
    # via google-auth
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via
    #   -r requirements.in
    #   cryptography
channels==3.0.5
    # via
    #   -r requirements.in
    #   channels-redis
    #   djangochannelsrestframework
channels-redis==3.4.1
    # via -r requirements.in
charset-normalizer==2.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via uvicorn
colorama==0.4.5
    # via
    #   click
    #   uvicorn
constantly==15.1.0
    # via twisted
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==37.0.4
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==3.0.2
    # via channels
deepgram-sdk==0.6.1
    # via -r requirements.in
django==4.1
    # via
    #   -r requirements.in
    #   channels
    #   django-debug-toolbar
    #   djangochannelsrestframework
    #   djangorestframework
    #   drf-yasg
django-debug-toolbar==3.2
    # via -r requirements.in
django-extensions==3.1.0
    # via -r requirements.in
django-redis-cache==3.0.0
    # via -r requirements.in
djangochannelsrestframework==1.1.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangochannelsrestframework
    #   drf-yasg
drf-yasg[validation]==1.21.4
    # via -r requirements.in
freezegun==1.2.1
    # via -r requirements.in
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
google-api-core[grpc]==2.10.1
    # via google-cloud-speech
google-auth==2.12.0
    # via google-api-core
google-cloud-speech==2.15.1
    # via -r requirements.in
googleapis-common-protos==1.56.4
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.49.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.49.1
    # via google-api-core
gunicorn==20.0.4
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
hiredis==2.0.0
    # via aioredis
httptools==0.4.0
    # via uvicorn
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.3
    # via
    #   anyio
    #   hyperlink
    #   requests
    #   twisted
    #   yarl
incremental==21.3.0
    # via twisted
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jsonschema==4.16.0
    # via swagger-spec-validator
markdown==3.4.1
    # via
    #   -r requirements.in
    #   markdown-urlize
markdown-urlize==0.2.0
    # via -r requirements.in
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
msgpack==1.0.4
    # via channels-redis
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
packaging==21.3
    # via drf-yasg
pillow==8.1.0
    # via -r requirements.in
proto-plus==1.22.1
    # via google-cloud-speech
protobuf==4.21.7
    # via
    #   google-api-core
    #   google-cloud-speech
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2==2.9.3
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
    #   service-identity
pyasn1-modules==0.2.8
    # via
    #   google-auth
    #   service-identity
pycparser==2.21
    # via cffi
pyopenssl==22.0.0
    # via
    #   -r requirements.in
    #   twisted
pyparsing==3.0.9
    # via
    #   -r requirements.in
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via freezegun
python-decouple==3.4
    # via -r requirements.in
python-dotenv==0.20.0
    # via uvicorn
pytz==2022.1
    # via
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0
    # via
    #   swagger-spec-validator
    #   uvicorn
redis==3.5.3
    # via
    #   -r requirements.in
    #   django-redis-cache
requests==2.28.1
    # via
    #   -r requirements.in
    #   coreapi
    #   google-api-core
rsa==4.9
    # via google-auth
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
sentry-sdk==0.19.5
    # via -r requirements.in
service-identity==21.1.0
    # via twisted
shortuuid==1.0.9
    # via -r requirements.in
six==1.15.0
    # via
    #   automat
    #   google-auth
    #   grpcio
    #   python-dateutil
    #   service-identity
    #   swagger-spec-validator
sniffio==1.2.0
    # via anyio
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
swagger-spec-validator==2.7.6
    # via drf-yasg
twisted[tls]==22.8.0
    # via daphne
txaio==22.2.1
    # via autobahn
typing-extensions==4.3.0
    # via twisted
tzdata==2022.4
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.2
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.18.2
    # via -r requirements.in
watchfiles==0.16.1
    # via uvicorn
watchgod==0.6
    # via -r requirements.in
websockets==10.3
    # via
    #   deepgram-sdk
    #   uvicorn
werkzeug==2.2.1
    # via -r requirements.in
whitenoise==5.2.0
    # via -r requirements.in
yarl==1.8.1
    # via aiohttp
zope-interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
